ORGANISING: moved all non-view handler functions into helpers.py to clean up the views.py file
